% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneSigR.R
\name{Export_ImmuneSigR_GMT}
\alias{Export_ImmuneSigR_GMT}
\alias{Export_GMT}
\title{Export the built-in GMT file}
\usage{
Export_ImmuneSigR_GMT(out_dir = ".", create_dir = TRUE)

Export_GMT(out_dir = ".", create_dir = TRUE)
}
\arguments{
\item{out_dir}{A character string specifying the output directory. Defaults to current working directory.}

\item{create_dir}{Logical. If TRUE, creates \code{out_dir} when it does not exist.}
}
\value{
Invisibly returns the exported GMT file path.
}
\description{
Copies the internal ImmuneSigR GMT database to a specified local directory.
}
\examples{
\dontrun{
Export_ImmuneSigR_GMT(out_dir = tempdir())
}
}
